move model into ngrx, clean up TODOs
