+++ src/app/downloads/downloads-chart.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 import * as moment from 'moment';
-import { EpisodeMetricsModel, PodcastMetricsModel, EpisodeModel, PodcastModel, FilterModel } from '../shared';
+import { EpisodeMetricsModel, PodcastMetricsModel, EpisodeModel, PodcastModel, FilterModel } from '../ngrx/model';
 import { TimeseriesChartModel, TimeseriesDatumModel } from 'ngx-prx-styleguide';
 
 @Component({
@@ -22,6 +22,7 @@ export class DownloadsChartComponent {
   podcastChartData: TimeseriesChartModel;
   chartData: TimeseriesChartModel[];
   // TODO: expand color support for more datasets
+  // --> episode selection
   colors = ['#000044', '#2C2C68', '#59598C', '#8686B0', '#B3B3D4'];
 
   constructor(private store: Store<any>) {
@@ -39,7 +40,6 @@ export class DownloadsChartComponent {
             if (this.podcastMetrics.length > 0) {
               this.podcastChartData = this.mapPodcastData(this.filter.podcast, this.podcastMetrics[0][metricsProperty + 'Others']);
               if (this.episodeChartData && this.episodeChartData.length > 0) {
-                // TODO: subtract episode from podcast data
                 this.chartData = [...this.episodeChartData, this.podcastChartData];
               } else {
                 this.chartData = [this.podcastChartData];
@@ -52,17 +52,16 @@ export class DownloadsChartComponent {
           if (!this.episodeMetricsStore) {
             this.episodeMetricsStore = this.store.select('episodeMetrics');
             this.episodeMetricsStore.subscribe((episodeMetrics: EpisodeMetricsModel[]) => {
-              this.episodeMetrics = episodeMetrics.filter((e: EpisodeMetricsModel) => {
-                // TODO: the hell is e/ep c'mon better names
-                return e.seriesId === this.filter.podcast.seriesId &&
-                  this.filter.episodes && this.filter.episodes.map(ep => ep.id).indexOf(e.id) !== -1;
+              this.episodeMetrics = episodeMetrics.filter((em: EpisodeMetricsModel) => {
+                return em.seriesId === this.filter.podcast.seriesId &&
+                  this.filter.episodes && this.filter.episodes.map(ef => ef.id).indexOf(em.id) !== -1;
               });
               if (this.episodeMetrics && this.episodeMetrics.length > 0) {
                 this.episodeChartData = this.episodeMetrics.map((episodeData: EpisodeMetricsModel) => {
                   const episode = this.filter.episodes.find(e => e.seriesId === episodeData.seriesId && e.id === episodeData.id);
                   return this.mapEpisodeData(episode, episodeData[metricsProperty]);
                 });
-                // sort these episode by their data total for the stacked chart
+                // sort these episodes by their data total for the stacked chart
                 this.episodeChartData.sort((a: TimeseriesChartModel, b: TimeseriesChartModel) => {
                   const getTotal = (data: TimeseriesDatumModel[]) => {
                     if (data.length > 0) {
@@ -78,12 +77,11 @@ export class DownloadsChartComponent {
                 // set the colors now that they are ordered
                 // TODO: can colors be optional to allow C3 to select colors
                 // TODO: chart component should not include color in C3 config when dataset is empty
+                // --> create styleguide tickets
                 for (let i = 0; i < this.episodeChartData.length; i++) {
                   this.episodeChartData[i].color = this.colors[i];
                 }
-                // TODO: sparse datasets need to be populated with zeroes for missing datetime entries
                 if (this.podcastChartData) {
-                  // TODO: subtract episode data from podcast data
                   this.chartData = [...this.episodeChartData, this.podcastChartData];
                 } else {
                   this.chartData = [...this.episodeChartData];
@@ -108,7 +106,6 @@ export class DownloadsChartComponent {
   }
 
   mapPodcastData(episode: PodcastModel, metrics: any[][]): TimeseriesChartModel {
-    // TODO this color does not stick on the podcast line because of the empty/sparse data sets
     return { data: this.mapData(metrics), label: episode.title, color: '#a3a3a3' };
   }
 }
