+++ src/app/ngrx/model/episode.model.ts
