+++ src/app/ngrx/model/filter.model.ts
