+++ src/app/ngrx/model/index.ts
@@ -1,0 +1,5 @@
+export { PodcastModel } from './podcast.model';
+export { EpisodeModel } from './episode.model';
+export { INTERVAL_DAILY, INTERVAL_HOURLY, INTERVAL_15MIN, IntervalModel,
+  PodcastMetricsModel, EpisodeMetricsModel, MetricsType } from './metrics.model';
+export { FilterModel } from './filter.model';
