+++ src/app/shared/index.ts
@@ -1,6 +1,2 @@
 export * from './shared.module';
-export { PodcastModel } from './model/podcast.model';
-export { EpisodeModel } from './model/episode.model';
-export { INTERVAL_DAILY, INTERVAL_HOURLY, INTERVAL_15MIN, IntervalModel,
-  PodcastMetricsModel, EpisodeMetricsModel } from './model/metrics.model';
-export { FilterModel } from './model/filter.model';
+
